Updated Hyperion Engine project and Toolchain setenv.bat script to use libyaul instead of SGL

+++ Projects/HyperionEngine/src/ECS/ECS.h
@@ -4,7 +4,7 @@
 
 #define ENTITY_CAPACITY 512
 
-typedef int EntityId;
+typedef short EntityId;
 
 typedef unsigned int ArchetypeId;
